+++ src/main/Game/Levels/Forest/FourWayForestFrame.cpp
@@ -6,13 +6,13 @@ FourWayForestFrame::FourWayForestFrame(Position referencePosition) : ForestFrame
     Position rightForestPosition = getStartingRightPositionFromWidth(forestWidthInTrees);
     auto finishedRightPosition = createDimensionedForest(forestWidthInTrees, forestHeightInTrees, rightForestPosition);
 
-    auto nextRightPosition = finishedRightPosition + Position(0, (float)treeHeight);
+    auto nextRightPosition = finishedRightPosition + Position(0, treeHeight);
     createDimensionedForest(forestWidthInTrees, forestHeightInTrees, nextRightPosition);
 
     Position leftForestPosition = {window->getTopLeftPosition()};
     auto finishedLeftPosition = createDimensionedForest(forestWidthInTrees,forestHeightInTrees, leftForestPosition);
 
-    auto nextLeftPosition = finishedLeftPosition + Position(0, (float)treeHeight);
+    auto nextLeftPosition = finishedLeftPosition + Position(0, treeHeight);
     createDimensionedForest(forestWidthInTrees, forestHeightInTrees, nextLeftPosition);
 
     openAt.openEverywhere();
